pyproject.toml
[tool.poetry]
name = "containers-sugar"
version = "1.0.0" # semantic-release
description = "Simplify the usage of containers"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "BSD 3 Clause"
include = [
  "*.cfg",
  "*.ini",
  "*.js",
  "*.json",
  "*.lock",
  "*.md",
  "*.py",
  "*.sh",
  "*.rst",
  "*.txt",
  "*.toml",
  "*.yml",
  "*.yaml",
  ".github",
  "conda",
  "docker",
  "docs",
  ".dockerignore",
  ".env.tpl",
  ".gitignore",
  "Makefile",
]
exclude = [
  ".git/*",
  ".env*",
]

[tool.poetry.scripts]
"containers-sugar" = "containers_sugar.__main__:app"


[tool.poetry.dependencies]
python = "^3.8"
podman-compose = "^1.0.3"
docker-compose = "^1.29.2"
sh = "^1.14.3"
pyyaml = "<6.0"
types-pyyaml = "^6.0.12.3"

[tool.poetry.dev-dependencies]
pytest = "^7"
blue = "^0.9.1"

isort = "^5"
pre-commit = "2"
flake8 = "^4"
mypy = "^0.931"
cookiecutter = "^1.7.3"
pytest-cov = "^3.0.0"
Jinja2 = "<3.1.0"
mkdocs = ">=1.2.3,<2"
mkdocs-exclude = ">=1.0.2,<2"
mkdocs-jupyter = ">=0.20.0,<1"
mkdocs-literate-nav = ">=0.4.1,<1"
mkdocs-macros-plugin = ">=0.6.3,<1"
mkdocs-material = ">=8.2.1,<9"
mkdocstrings = ">=0.17.0,<0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.blue]
line-length = 79  # this is the default

[tool.isort]
ensure_newline_before_comments = true
line_length = 79
multi_line_output = 3
include_trailing_comma = true
skip_glob = ["docs/*", "*.egg-info"]

[tool.mypy]
ignore_missing_imports = true
